+++ server/src/models.rs
@@ -5,7 +5,7 @@ use crate::schema::users;
 #[derive(Serialize, Deserialize, Clone, Queryable, Debug, Insertable, AsChangeset)]
 #[table_name = "users"]
 pub struct User {
-    pub uid: Option<i64>,
+    pub uid: Option<i32>,
     pub name: String,
     pub login: String,
     pub hashword: String,
